refactor(models): moved models into a submodule
