+++ clientforge/models/results.py
@@ -1,238 +1,24 @@
-"""Models for the results of the API requests.
+"""A module for the results of the models."""
 
-The ForgeModel is designed to be a base class for all models defined in the
-user's code. The Response class is used to represent the response from the
-server and provides methods to convert the response to a model or get values
-from the JSON content.
-"""
+from __future__ import annotations
 
 import json
 from abc import ABCMeta
 from copy import deepcopy
 from dataclasses import dataclass
-from typing import Generic, TypeVar
+from typing import TYPE_CHECKING, Generic, TypeVar
 
 import httpx
 import jsonpath_ng.ext
 from dataclass_wizard import JSONWizard
 
 from clientforge.exceptions import InvalidJSONResponse
+from clientforge.models.fields import Field
+
+if TYPE_CHECKING:
+    from clientforge.models.fields import Condition
 
 MODEL = TypeVar("MODEL", bound="type[ForgeModel]")
-
-
-class ConditionOperator:
-    """An enum to represent the operators for a condition."""
-
-    LT = "lt"
-    LE = "le"
-    EQ = "eq"
-    GE = "ge"
-    GT = "gt"
-    LEN_LT = "len_lt"
-    LEN_LE = "len_le"
-    LEN_EQ = "len_eq"
-    LEN_GE = "len_ge"
-    LEN_GT = "len_gt"
-
-
-class FieldLength:
-    """A class to represent an operation on a field in a model."""
-
-    def __init__(self, field: "Field") -> None:
-        """Initialize the field length.
-
-        Parameters
-        ----------
-            field: Field
-                The field to apply the condition to.
-        """
-        self.field = field
-
-    def __lt__(self, other):
-        """Return a condition for the length being lt the other value."""
-        return Condition(self.field, ConditionOperator.LEN_LT, other)
-
-    def __le__(self, other):
-        """Return a condition for the length being lt or equal to the other value."""
-        return Condition(self.field, ConditionOperator.LEN_LE, other)
-
-    def __eq__(self, other):
-        """Return a condition for the length being equal to the other value."""
-        return Condition(self.field, ConditionOperator.LEN_EQ, other)
-
-    def __ge__(self, other):
-        """Return a condition for the length being gt or equal to the other value."""
-        return Condition(self.field, ConditionOperator.LEN_GE, other)
-
-    def __gt__(self, other):
-        """Return a condition for the length being gt the other value."""
-        return Condition(self.field, ConditionOperator.LEN_GT, other)
-
-    def __str__(self):
-        """Return a string representation of the field."""
-        return f"FieldLength({self.field})"
-
-
-class FieldIterable:
-    """A class to represent an iterable field in a model."""
-
-    def __init__(self, field: "Field") -> None:
-        """Initialize the field iterable.
-
-        Parameters
-        ----------
-            field: Field
-                The field to apply the condition to.
-        """
-        self.field = field
-
-    def any(self, condition: "Condition"):
-        """Return a condition for any items in the iterable field."""
-        return ConditionIterable(self.field, condition, False)
-
-    def all(self, condition: "Condition"):
-        """Return a condition for all items in the iterable field."""
-        return ConditionIterable(self.field, condition, True)
-
-    def __str__(self):
-        """Return a string representation of the field."""
-        return f"FieldIterable({self.field})"
-
-
-class Field:
-    """A class to represent a field in a model."""
-
-    @property
-    def length(self):
-        """Return a FieldLength object for the field."""
-        return FieldLength(self)
-
-    @property
-    def where(self):
-        """Return a FieldIterable object for the field."""
-        return FieldIterable(self)
-
-    def __set_name__(self, owner, name) -> None:
-        """Set the name of the field and the owner class."""
-        self.owner: type[ForgeModel] = owner
-        self.name: str = name
-
-    def __lt__(self, other):
-        """Return a condition for the field being lt the other value."""
-        return Condition(self, ConditionOperator.LT, other)
-
-    def __le__(self, other):
-        """Return a condition for the field being lt or equal to the other value."""
-        return Condition(self, ConditionOperator.LE, other)
-
-    def __eq__(self, other):
-        """Return a condition for the field being equal to the other value."""
-        return Condition(self, ConditionOperator.EQ, other)
-
-    def __ge__(self, other):
-        """Return a condition for the field being gt or equal to the other value."""
-        return Condition(self, ConditionOperator.GE, other)
-
-    def __gt__(self, other):
-        """Return a condition for the field being gt the other value."""
-        return Condition(self, ConditionOperator.GT, other)
-
-    def __hash__(self):
-        """Return a hash of the field name (required for dataclasses)."""
-        return hash(self.name)
-
-    def __str__(self):
-        """Return a string representation of the field."""
-        return f"Field({self.owner.__name__}.{self.name})"
-
-
-class ConditionIterable:
-    """A class to represent a condition on an iterable field in a model."""
-
-    def __init__(self, field: Field, condition: "Condition", strict: bool) -> None:
-        """Initialize the condition.
-
-        Parameters
-        ----------
-            field: Field
-                The field to apply the condition to.
-            condition: Condition
-                The condition to apply to the field.
-            strict: bool
-                Whether to apply the condition to all or any of the items.
-        """
-        self.field = field
-        self.condition = condition
-        self.strict = strict
-
-    def evaluate(self, model: "ForgeModel") -> bool:
-        """Evaluate the condition on the model."""
-        field_value: list[ForgeModel] = getattr(model, self.field.name)
-
-        if self.strict:
-            out = all(self.condition.evaluate(item) for item in field_value)
-        else:
-            out = any(self.condition.evaluate(item) for item in field_value)
-
-        return out
-
-    def __str__(self):
-        """Return a string representation of the condition."""
-        return f"ConditionIterable({self.condition} over {self.field})"
-
-
-class Condition:
-    """A class to represent a condition on a model field."""
-
-    def __init__(self, field: Field, operator: ConditionOperator, value) -> None:
-        """Initialize the condition.
-
-        Parameters
-        ----------
-            field: Field
-                The field to apply the condition to.
-            operator: ConditionOperator
-                The operator to apply to the field.
-            value: Any
-                The value to compare the field to.
-        """
-        self.field = field
-        self.operator = operator
-        self.value = value
-
-    def evaluate(self, model: "ForgeModel") -> bool:
-        """Evaluate the condition on the model."""
-        field_value = getattr(model, self.field.name)
-
-        match self.operator:
-            case ConditionOperator.LT:
-                return field_value < self.value
-            case ConditionOperator.LE:
-                return field_value <= self.value
-            case ConditionOperator.EQ:
-                return field_value == self.value
-            case ConditionOperator.GE:
-                return field_value >= self.value
-            case ConditionOperator.GT:
-                return field_value > self.value
-
-            case ConditionOperator.LEN_LT:
-                return len(field_value) < self.value
-            case ConditionOperator.LEN_LE:
-                return len(field_value) <= self.value
-            case ConditionOperator.LEN_EQ:
-                return len(field_value) == self.value
-            case ConditionOperator.LEN_GE:
-                return len(field_value) >= self.value
-            case ConditionOperator.LEN_GT:
-                return len(field_value) > self.value
-            case _:
-                raise ValueError(f"Unsupported operator: {self.operator}")
-
-    def __str__(self) -> str:
-        """Return a string representation of the condition."""
-        return f"Condition({self.field} {self.operator} {self.value})"
 
 
 class Result(Generic[MODEL]):
@@ -255,7 +41,7 @@ class Result(Generic[MODEL]):
 
         self.model = model
 
-    def filter(self, *conditions: Condition) -> "Result":
+    def filter(self, *conditions: Condition) -> Result:
         """Filter the results based on the conditions.
 
         Parameters
@@ -297,7 +83,7 @@ class Result(Generic[MODEL]):
             if all(cond.evaluate(item) for cond in conditions)
         )
 
-    def query(self, query: str | jsonpath_ng.JSONPath) -> "Result":
+    def query(self, query: str | jsonpath_ng.JSONPath) -> Result:
         """Filter the results using a JSONPath query.
 
         Parameters
